Refactor process and send flight to studio
